# lib/api-zod/src/generated/types/invoicesListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arabic POS System API
 * OpenAPI spec version: 0.1.0
 */
import type { Invoice } from "./invoice";
import type { PaginationMeta } from "./paginationMeta";

export interface InvoicesListResponse {
  data: Invoice[];
  meta: PaginationMeta;
}